refactor: extract commands into seperate module
