+++ src/webdynpro/command/element/mod.rs
@@ -1,0 +1,5 @@
+/// 액션 분류의 엘리먼트([`Button`], [`Link`])를 위한 명령
+pub mod action;
+
+/// 선택 분류의 엘리먼트([`ListBox`], [`ComboBox`])를 위한 명령
+pub mod selection;
